10.04
all is good except fivedaysJson is null....

+++ app/src/main/java/com/dmitrysukhov/weatherapp/model/wrappers/CurrentWeatherWrapper.java
@@ -165,7 +165,7 @@ public class CurrentWeatherWrapper {
 
     public static class MetricSpeedOfWind {
         @SerializedName("Value")
-        private int speedOfWindValue;
+        private double speedOfWindValue;
         @SerializedName("Unit")
         private String unitSpeedOfWind;
